+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.1
 * `wps dotfiles`, backup and restore dotfiles
 * `wps zstds`, use tar and zstd to compress and decompress directory
+* `starter workspace` ,add funsapce and acespace Coderepo worksapce
 
 ## v0.1.0
 * starter feature ,initial project, add feature to project.